Relocating organization classes to organization package
