+++ src/main/java/com/ninjaone/dundie_awards/DataLoader.java
@@ -2,8 +2,8 @@ package com.ninjaone.dundie_awards;
 
 import com.ninjaone.dundie_awards.employee.Employee;
 import com.ninjaone.dundie_awards.employee.EmployeeRepository;
-import com.ninjaone.dundie_awards.model.Organization;
-import com.ninjaone.dundie_awards.repository.OrganizationRepository;
+import com.ninjaone.dundie_awards.organization.Organization;
+import com.ninjaone.dundie_awards.organization.OrganizationRepository;
 import org.springframework.boot.CommandLineRunner;
 import org.springframework.stereotype.Component;
