+++ src/main/java/com/ninjaone/dundie_awards/organization/OrganizationRepository.java
@@ -1,6 +1,5 @@
-package com.ninjaone.dundie_awards.repository;
+package com.ninjaone.dundie_awards.organization;
 
-import com.ninjaone.dundie_awards.model.Organization;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
